Add user without authentication

+++ api/src/main/java/com/moneymanager/api/dtos/BankRecordDetailsDto.java
@@ -1,18 +1,22 @@
 package com.moneymanager.api.dtos;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.util.List;
+/**
+ * BankRecordDetailsDto is returned in place of BankRecord
+ * It contains linked financial transactions and file attachments
+ */
 
 @Data
 @AllArgsConstructor
 public class BankRecordDetailsDto {
     private Long id;
-    private Integer year;
-    private Integer month;
-    private Integer day;
-    private Double amount;
+    private Short year;
+    private Byte month;
+    private Byte day;
+    private Long amount;
     private String name;
     private List<FinancialTransactionDto> financialTransactions;
     private List<FileAttachmentDto> fileAttachments;
