+++ api/src/main/java/com/moneymanager/api/dtos/DateAmountDto.java
@@ -3,12 +3,16 @@ package com.moneymanager.api.dtos;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+/**
+ * DateAmountDto is returned in place of DateAmount
+ */
+
 @Data
 @AllArgsConstructor
 public class DateAmountDto {
     private Long id;
-    private Integer year;
-    private Integer month;
-    private Integer day;
-    private Double amount;
+    private Short year;
+    private Byte month;
+    private Byte day;
+    private Long amount;
 }
