+++ api/src/main/java/com/moneymanager/api/dtos/FinancialTransactionDetailsDto.java
@@ -1,18 +1,22 @@
 package com.moneymanager.api.dtos;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.util.List;
+/**
+ * FinancialTransactionDetailsDto is returned in place of FinancialTransactionDetails
+ * It contains linked bank records and file attachments
+ */
 
 @Data
 @AllArgsConstructor
 public class FinancialTransactionDetailsDto {
     private Long id;
-    private Integer year;
-    private Integer month;
-    private Integer day;
-    private Double amount;
+    private Short year;
+    private Byte month;
+    private Byte day;
+    private Long amount;
     private String name;
     private List<BankRecordDto> bankRecords;
     private List<FileAttachmentDto> fileAttachments;
