+++ api/src/main/java/com/moneymanager/api/dtos/FinancialTransactionDto.java
@@ -3,13 +3,18 @@ package com.moneymanager.api.dtos;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+/**
+ * FinancialTransactionDto is returned in place of FinancialTransaction
+ * It does not contain any linked items
+ */
+
 @Data
 @AllArgsConstructor
 public class FinancialTransactionDto {
     private Long id;
-    private Integer year;
-    private Integer month;
-    private Integer day;
-    private Double amount;
+    private Short year;
+    private Byte month;
+    private Byte day;
+    private Long amount;
     private String name;
 }
