+++ api/src/main/java/com/moneymanager/api/requests/BankRecordRequest.java
@@ -2,11 +2,15 @@ package com.moneymanager.api.requests;
 
 import lombok.Data;
 
+/**
+ * A BankRecordRequest is used to create a BankRecord
+ */
+
 @Data
 public class BankRecordRequest {
-    private Integer year;
-    private Integer month;
-    private Integer day;
-    private Double amount;
+    private Short year;
+    private Byte month;
+    private Byte day;
+    private Long amount;
     private String name;
 }
