+++ api/src/main/java/com/moneymanager/api/requests/DateAmountCreateRequest.java
@@ -2,10 +2,14 @@ package com.moneymanager.api.requests;
 
 import lombok.Data;
 
+/**
+ * A DateAmountCreate request is used to create a DateAmount
+ */
+
 @Data
 public class DateAmountCreateRequest {
-    private Integer year;
-    private Integer month;
-    private Integer day;
-    private Double amount;
+    private Short year;
+    private Byte month;
+    private Byte day;
+    private Long amount;
 }
