+++ api/src/main/java/com/moneymanager/api/requests/DateAmountUpdateRequest.java
@@ -2,7 +2,11 @@ package com.moneymanager.api.requests;
 
 import lombok.Data;
 
+/**
+ * A DateAmountUpdateRequest is used to update a DateRequest
+ */
+
 @Data
 public class DateAmountUpdateRequest {
-    private Double amount;
+    private Long amount;
 }
